+++ eslint.config.mjs
@@ -64,7 +64,7 @@ const eslintConfig = defineConfig([
       "no-shadow": "off",
 
       /* 🔟 Max line length */
-      // "max-lines": ["error", { max: 180 }],
+      "max-lines": ["error", { max: 180 }],
     },
   },
