--- src/app/_components/editor/live-chart-preview.tsx
@@ -1,299 +1,0 @@
-"use client";
-
-import { motion } from "framer-motion";
-import {
-  BarChart,
-  Bar,
-  LineChart,
-  Line,
-  AreaChart,
-  Area,
-  PieChart,
-  Pie,
-  ScatterChart,
-  Scatter,
-  XAxis,
-  YAxis,
-  CartesianGrid,
-  Tooltip,
-  ResponsiveContainer,
-  Cell,
-  Legend,
-} from "recharts";
-import { BarChart3 } from "lucide-react";
-import type { ChartSuggestion } from "./chart-suggestions";
-import type { UploadedFile } from "./excel-upload";
-
-type LiveChartPreviewProps = {
-  file: UploadedFile | null;
-  suggestion: ChartSuggestion | null;
-  isLoading?: boolean;
-};
-
-const COLORS = [
-  "hsl(var(--chart-1))",
-  "hsl(var(--chart-2))",
-  "hsl(var(--chart-3))",
-  "hsl(var(--chart-4))",
-  "hsl(var(--chart-5))",
-];
-
-export function LiveChartPreview({
-  file,
-  suggestion,
-  isLoading,
-}: LiveChartPreviewProps) {
-  if (isLoading) {
-    return (
-      <div className="h-full flex items-center justify-center">
-        <motion.div
-          initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
-          className="text-center"
-        >
-          <div className="relative w-20 h-20 mx-auto mb-4">
-            <motion.div className="absolute inset-0 rounded-full border-4 border-primary/20" />
-            <motion.div
-              className="absolute inset-0 rounded-full border-4 border-primary border-t-transparent"
-              animate={{ rotate: 360 }}
-              transition={{ duration: 1, repeat: Infinity, ease: "linear" }}
-            />
-          </div>
-          <p className="text-muted-foreground">Generating chart...</p>
-        </motion.div>
-      </div>
-    );
-  }
-
-  if (!file || !suggestion || !suggestion.xAxis || !suggestion.yAxis) {
-    return (
-      <div className="h-full flex items-center justify-center">
-        <motion.div
-          initial={{ opacity: 0, y: 20 }}
-          animate={{ opacity: 1, y: 0 }}
-          className="text-center max-w-md"
-        >
-          <div className="p-6 rounded-2xl bg-muted/30 inline-block mb-4">
-            <BarChart3 className="h-12 w-12 text-muted-foreground" />
-          </div>
-          <h3 className="text-xl font-semibold mb-2">Chart Preview</h3>
-          <p className="text-muted-foreground">
-            {!file
-              ? "Upload an Excel file and select a chart type to see your data visualized here"
-              : "Select a chart suggestion and configure axes to preview your chart"}
-          </p>
-        </motion.div>
-      </div>
-    );
-  }
-
-  // Prepare chart data
-  const chartData = file.data.slice(0, 50).map((row) => ({
-    x: row[suggestion.xAxis!],
-    y: Number(row[suggestion.yAxis!]) || 0,
-    name: String(row[suggestion.xAxis!]),
-    value: Number(row[suggestion.yAxis!]) || 0,
-  }));
-
-  const renderChart = () => {
-    const commonProps = {
-      data: chartData,
-      margin: { top: 20, right: 30, left: 20, bottom: 60 },
-    };
-
-    switch (suggestion.type) {
-      case "bar":
-        return (
-          <BarChart {...commonProps}>
-            <CartesianGrid strokeDasharray="3 3" stroke="hsl(var(--border))" />
-            <XAxis
-              dataKey="name"
-              stroke="hsl(var(--muted-foreground))"
-              tick={{ fill: "hsl(var(--muted-foreground))", fontSize: 12 }}
-              angle={-45}
-              textAnchor="end"
-              height={60}
-            />
-            <YAxis
-              stroke="hsl(var(--muted-foreground))"
-              tick={{ fill: "hsl(var(--muted-foreground))", fontSize: 12 }}
-            />
-            <Tooltip
-              contentStyle={{
-                backgroundColor: "hsl(var(--card))",
-                border: "1px solid hsl(var(--border))",
-                borderRadius: "8px",
-              }}
-            />
-            <Bar dataKey="value" radius={[4, 4, 0, 0]}>
-              {chartData.map((_, index) => (
-                <Cell
-                  key={`cell-${index}`}
-                  fill={COLORS[index % COLORS.length]}
-                />
-              ))}
-            </Bar>
-          </BarChart>
-        );
-
-      case "line":
-        return (
-          <LineChart {...commonProps}>
-            <CartesianGrid strokeDasharray="3 3" stroke="hsl(var(--border))" />
-            <XAxis
-              dataKey="name"
-              stroke="hsl(var(--muted-foreground))"
-              tick={{ fill: "hsl(var(--muted-foreground))", fontSize: 12 }}
-              angle={-45}
-              textAnchor="end"
-              height={60}
-            />
-            <YAxis
-              stroke="hsl(var(--muted-foreground))"
-              tick={{ fill: "hsl(var(--muted-foreground))", fontSize: 12 }}
-            />
-            <Tooltip
-              contentStyle={{
-                backgroundColor: "hsl(var(--card))",
-                border: "1px solid hsl(var(--border))",
-                borderRadius: "8px",
-              }}
-            />
-            <Line
-              type="monotone"
-              dataKey="value"
-              stroke={COLORS[1]}
-              strokeWidth={2}
-              dot={{ fill: COLORS[1], strokeWidth: 2 }}
-              activeDot={{ r: 6, fill: COLORS[1] }}
-            />
-          </LineChart>
-        );
-
-      case "area":
-        return (
-          <AreaChart {...commonProps}>
-            <defs>
-              <linearGradient id="areaGradient" x1="0" y1="0" x2="0" y2="1">
-                <stop offset="5%" stopColor={COLORS[2]} stopOpacity={0.8} />
-                <stop offset="95%" stopColor={COLORS[2]} stopOpacity={0} />
-              </linearGradient>
-            </defs>
-            <CartesianGrid strokeDasharray="3 3" stroke="hsl(var(--border))" />
-            <XAxis
-              dataKey="name"
-              stroke="hsl(var(--muted-foreground))"
-              tick={{ fill: "hsl(var(--muted-foreground))", fontSize: 12 }}
-              angle={-45}
-              textAnchor="end"
-              height={60}
-            />
-            <YAxis
-              stroke="hsl(var(--muted-foreground))"
-              tick={{ fill: "hsl(var(--muted-foreground))", fontSize: 12 }}
-            />
-            <Tooltip
-              contentStyle={{
-                backgroundColor: "hsl(var(--card))",
-                border: "1px solid hsl(var(--border))",
-                borderRadius: "8px",
-              }}
-            />
-            <Area
-              type="monotone"
-              dataKey="value"
-              stroke={COLORS[2]}
-              fillOpacity={1}
-              fill="url(#areaGradient)"
-            />
-          </AreaChart>
-        );
-
-      case "pie":
-        return (
-          <PieChart>
-            <Pie
-              data={chartData.slice(0, 8)}
-              cx="50%"
-              cy="50%"
-              labelLine={false}
-              label={({ name, percent }) =>
-                `${name}: ${((percent ?? 0) * 100).toFixed(0)}%`
-              }
-              outerRadius={120}
-              fill="#8884d8"
-              dataKey="value"
-            >
-              {chartData.slice(0, 8).map((_, index) => (
-                <Cell
-                  key={`cell-${index}`}
-                  fill={COLORS[index % COLORS.length]}
-                />
-              ))}
-            </Pie>
-            <Tooltip
-              contentStyle={{
-                backgroundColor: "hsl(var(--card))",
-                border: "1px solid hsl(var(--border))",
-                borderRadius: "8px",
-              }}
-            />
-            <Legend />
-          </PieChart>
-        );
-
-      case "scatter":
-        return (
-          <ScatterChart {...commonProps}>
-            <CartesianGrid strokeDasharray="3 3" stroke="hsl(var(--border))" />
-            <XAxis
-              type="category"
-              dataKey="name"
-              stroke="hsl(var(--muted-foreground))"
-              tick={{ fill: "hsl(var(--muted-foreground))", fontSize: 12 }}
-            />
-            <YAxis
-              type="number"
-              dataKey="value"
-              stroke="hsl(var(--muted-foreground))"
-              tick={{ fill: "hsl(var(--muted-foreground))", fontSize: 12 }}
-            />
-            <Tooltip
-              contentStyle={{
-                backgroundColor: "hsl(var(--card))",
-                border: "1px solid hsl(var(--border))",
-                borderRadius: "8px",
-              }}
-            />
-            <Scatter name="Data" data={chartData} fill={COLORS[4]} />
-          </ScatterChart>
-        );
-
-      default:
-        return null;
-    }
-  };
-
-  return (
-    <motion.div
-      initial={{ opacity: 0 }}
-      animate={{ opacity: 1 }}
-      className="h-full flex flex-col"
-    >
-      {/* Chart Title */}
-      <div className="p-4 border-b border-border/50">
-        <h3 className="font-semibold">{suggestion.title}</h3>
-        <p className="text-sm text-muted-foreground">
-          {suggestion.xAxis} vs {suggestion.yAxis}
-        </p>
-      </div>
-
-      {/* Chart */}
-      <div className="flex-1 p-4">
-        <ResponsiveContainer width="100%" height="100%">
-          {renderChart()}
-        </ResponsiveContainer>
-      </div>
-    </motion.div>
-  );
-}
